auth-service/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
sqlmodel>=0.0.8
sqlalchemy>=1.4.0,<2.0.0
alembic>=1.7.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
argon2-cffi
firebase-admin
python-dotenv
pytest
httpx
google-auth 
pydantic[email]